requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
selenium==4.15.2
beautifulsoup4==4.12.2
webdriver-manager==4.0.1
sendgrid==6.10.0
pydantic==2.4.2
python-multipart==0.0.6
jinja2==3.1.2 